% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnlp.r
\name{plotnlp}
\alias{plotnlp}
\title{plotnlp
Draw the barplot of negative log10 p-values for coregulated transcription factors}
\usage{
custom<-c("Hoxa10","Hoxa11","Hoxa13","Hoxa13","Hoxa2","Hoxa5","Hoxa9","Hoxb13","Hoxb4","Hoxb8","Hoxc13","Hoxd12","Hoxd13","Ikbkb")

res<-tfcrcalc(custom)

plotnlp(res)
}
\arguments{
\item{df}{a dataframe of results obtained with tfcrcalc function}
}
\description{
plotnlp
Draw the barplot of negative log10 p-values for coregulated transcription factors
}
\examples{
custom<-c("Hoxa10","Hoxa11","Hoxa13","Hoxa13","Hoxa2","Hoxa5","Hoxa9","Hoxb13","Hoxb4","Hoxb8","Hoxc13","Hoxd12","Hoxd13","Ikbkb")
res<-tfcrcalc(custom)
res<-head(res,n=15)
plotnlp(res)
}
