% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcrcalc.r
\name{tfcrcalc}
\alias{tfcrcalc}
\title{tfcrcalc
Calculates transcription factor coregulation enrichment on the input mice gene list}
\usage{
custom<-c("Hoxa10","Hoxa11","Hoxa13","Hoxa13","Hoxa2","Hoxa5","Hoxa9","Hoxb13","Hoxb4","Hoxb8","Hoxc13","Hoxd12","Hoxd13","Ikbkb")

res<-tfcrcalc(custom)
}
\arguments{
\item{custom}{a vector of mice gene symbol containing possible transcription factors}
}
\description{
tfcrcalc
Calculates transcription factor coregulation enrichment on the input mice gene list
}
\examples{
custom<-c("Hoxa10","Hoxa11","Hoxa13","Hoxa13","Hoxa2","Hoxa5","Hoxa9","Hoxb13","Hoxb4","Hoxb8","Hoxc13","Hoxd12","Hoxd13","Ikbkb")
res<-tfcrcalc(custom)
res<-head(res,n=15)
res
}
