% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcrnet.R
\name{tfcrnet}
\alias{tfcrnet}
\title{tfcrnet
built the network of tf coenriched}
\usage{
custom<-c("Hoxa10","Hoxa11","Hoxa13","Hoxa13","Hoxa2","Hoxa5","Hoxa9","Hoxb13","Hoxb4","Hoxb8","Hoxc13","Hoxd12","Hoxd13","Ikbkb")

res<-tfcrcalc(custom)

tfcrnet(custom,res,layout=layout_as_star,cex=1,distance=2)
}
\arguments{
\item{custom}{a vector of mice gene symbol containing possible tf}

\item{res}{a dataframe of results obtained with tfcrcalc function}

\item{layout}{string of layout type for the network (layout_as_star, layout_components, layout_in_circle, layout_nicely,layout_on_grid, layout_on_sphere, layout_randomly, layout_with_dh, layout_with_drl, layout_with_fr, layout_with_gem, layout_with_graphopt, layout_with_kk, layout_with_lgl, layout_with_mds)}

\item{cex}{size of the vertex labels}

\item{distance}{distance between vertex and its label}
}
\description{
tfcrnet
built the network of tf coenriched
}
\examples{
custom<-c("Hoxa10","Hoxa11","Hoxa13","Hoxa13","Hoxa2","Hoxa5","Hoxa9","Hoxb13","Hoxb4","Hoxb8","Hoxc13","Hoxd12","Hoxd13","Ikbkb")
res<-tfcrcalc(custom)
res<-head(res,n=5)
res
tfcrnet(custom,res,layout=layout_nicely,cex=1,distance=1.5)
tfcrnet(custom,res,layout=layout_as_star,cex=1,distance=1.5)
}
